NF: UndoableReview as anonymous class
